cloudinary added

+++ src/controllers/productController.js
@@ -1,18 +1,24 @@
 import { successResponse } from "../helper/response.js";
 import Product from "../model/productModel.js";
+import { uploadImage } from "../helper/cloudinaryUpload.js";
+import cloudinary from "../config/cloudinary.js";
 
 const createProduct = async (req, res, next)=>{
 
     const { productName } = req.body;
    
     try {
-        let productObj={ productName };
+        
         console.log(req.file + "fi")
-        if(req.file){
-             productObj. productImg  = `${req.protocol}://${req.get("host")}/uploads/${req.file.filename}`;
-        }
+        
+       const result=await uploadImage(req.file.path);
+        
 
-       const newProduct =  await Product.create(productObj);
+       const newProduct =  await Product.create({
+            productName: productName,
+            productImg: result.secure_url,
+            publicId: result.public_id
+        });
 
         return successResponse(res,{
             statusCode  : 201,
@@ -42,9 +48,25 @@ const updateProduct = async (req, res, next)=>{
         const { id } = req.params;
         const { productName } = req.body;
         try {
+
+       const product = await Product.findById(id);
+
+        if(!product){
+
+            return res.status(404).json({
+                message:"Not found"
+            });
+
+        }
+
+
+
             let productObj={ productName };
             if(req.file){
-                 productObj. productImg  = `${req.protocol}://${req.get("host")}/uploads/${req.file.filename}`;
+                await cloudinary.uploader.destroy(product.publicId);
+                const result = await uploadImage(req.file.path);
+                productObj.productImg = result.secure_url;
+                productObj.publicId = result.public_id;
             }
             const updatedProduct = await Product.findByIdAndUpdate(id, productObj, { new: true });
             return successResponse(res,{
@@ -60,6 +82,17 @@ const updateProduct = async (req, res, next)=>{
 const deleteProduct = async (req, res, next)=>{
     const { id } = req.params;
     try {
+
+        const product = await Product.findById(id);
+
+        if(!product){
+
+            return res.status(404).json({
+                message:"Not found"
+            });
+
+        }
+        await cloudinary.uploader.destroy(product.publicId);
         await Product.findByIdAndDelete(id);
         return successResponse(res,{
             statusCode  : 200,
